fe/requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
django==4.1.3
idna==3.4 ; python_version >= '3.5'
psycopg2-binary==2.9.5
requests==2.28.1 ; python_version >= '3.0'
sqlparse==0.4.3 ; python_version >= '3.5'
stripe==5.0.0
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
phonenumbers
django-phonenumber-field
requests
django-mathfilters
